feat: initial neovim config revision
I'll go through what each file is.

init.lua: this sources every other file. Currently those files are
explicitly named, and note that `plugin_manager.lua` is not included as
I'm still working on my own plugin manager and it's not ready yet.

The files in question that are explicitly sourced:

 - autocmd.lua
 - keymaps.lua
 - lsp.lua
 - options.lua
 - plugin_manager.lua

autocmd.lua

The main purpose of this file is to regulate the lights and darks of the
vim color scheme that I use (namely, 'vim').

keymaps.lua

Self explanatory. Contains non-filetype-specific keymaps. Note that LSP
keymaps are not included and are actually in lsp.lua

lsp.lua

Contains LSP key maps and auto commands that are responsible for starting
filetype-specific language servers.

options.lua

Contains global options, sorted into a table for better readability.